Properly generating and uploading a public (signature verification) key.
